% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perma_link.R
\name{perma_assert_addin_requirements}
\alias{perma_assert_addin_requirements}
\title{`perma_assert_addin_requirements`}
\usage{
perma_assert_addin_requirements()
}
\description{
`perma_assert_addin_requirements`
}
