% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perma_link.R
\name{perma_check_files}
\alias{perma_check_files}
\title{`perma_check_file_state`}
\usage{
perma_check_files(file_linked)
}
\arguments{
\item{file_linked}{check file linked for uncomitted work or comitted work
that is not pushed. This will throw a warning if the file does not match the
remote.}
}
\description{
`perma_check_file_state`
}
