% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perma_link.R
\name{perma_get_document_selection}
\alias{perma_get_document_selection}
\title{`perma_get_document_selection`}
\usage{
perma_get_document_selection(editor_location)
}
\arguments{
\item{editor_location}{result of rstudioapi::getSourceEditorContext()}
}
\description{
`perma_get_document_selection`
}
