% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perma_link.R
\name{perma_navigate_to_git_link_state}
\alias{perma_navigate_to_git_link_state}
\title{`perma_navigate_to_git_link_state`}
\usage{
perma_navigate_to_git_link_state(perma_link_info)
}
\arguments{
\item{perma_link_info}{item that contains parsed perma link issues}
}
\description{
`perma_navigate_to_git_link_state`
}
