% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perma_link.R
\name{perma_navigate_to_link}
\alias{perma_navigate_to_link}
\title{`perma_navigate_to_link_spot`}
\usage{
perma_navigate_to_link(link_items)
}
\arguments{
\item{link_items}{item that contains parsed perma link issues}
}
\description{
`perma_navigate_to_link_spot`
}
