% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perma_link.R
\name{perma_parse_and_validate_link}
\alias{perma_parse_and_validate_link}
\title{`perma_parse_and_validate_link`
This funicton will crash if we do not have a valid github perma link}
\usage{
perma_parse_and_validate_link(link, host = "https://github.com/")
}
\arguments{
\item{link}{perma link}

\item{host}{hose name (DEFAULT to https://github.com)}
}
\description{
`perma_parse_and_validate_link`
This funicton will crash if we do not have a valid github perma link
}
